Exception handling - delete

+++ src/main/java/com/springproject/course/services/UserService.java
@@ -4,36 +4,49 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.springproject.course.entities.User;
 import com.springproject.course.repositories.UserRepository;
+import com.springproject.course.resources.exceptions.DatabaseException;
 import com.springproject.course.services.exceptions.ResourceNotFoundException;
 
-@Service //Marcado no Registro de componentes, para poder ser ligado como dependência por outras classes
+@Service // Marcado no Registro de componentes, para poder ser ligado como dependência
+			// por outras classes
 //como no Autowired do UserRespository
 public class UserService {
-	
+
 	@Autowired
 	UserRepository userRepository;
 
-	public List<User> findAll(){
+	public List<User> findAll() {
 		return userRepository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = userRepository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User user) {
 		return userRepository.save(user);
 	}
-	
+
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {
+			if (!userRepository.existsById(id)) {
+				throw new EmptyResultDataAccessException(0);
+			}
+			userRepository.deleteById(id);
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
+
 	public User update(Long id, User user) {
 		User entity = userRepository.findById(id).get();
 		updateData(entity, user);
